refactor: RacingMager와 RacingResult 패키지 이동
