+++ src/main/java/racingcar/Application.java
@@ -1,8 +1,6 @@
 package racingcar;
 
 import racingcar.domain.Participants;
-import racingcar.domain.RacingManager;
-import racingcar.domain.RacingResult;
 import racingcar.domain.RandomSupplier;
 import racingcar.domain.Round;
 import racingcar.view.InputView;
