+++ src/test/java/racingcar/domain/RacingManagerTest.java
@@ -5,6 +5,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import racingcar.RacingManager;
+import racingcar.RacingResult;
 
 class RacingManagerTest {
