+++ src/test/java/racingcar/domain/RacingResultTest.java
@@ -5,6 +5,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import racingcar.RacingResult;
 
 class RacingResultTest {
